Moved real traffic controller to separate module
